refactor(core): move wire-format stream codec out of provider into core
