+++ crates/core/Cargo.toml
@@ -17,6 +17,7 @@ sonic = ["dep:sonic-rs"]
 [dependencies]
 # crates-io
 bytes.workspace = true
+futures.workspace = true
 futures-core.workspace = true
 rand.workspace = true
 serde.workspace = true
