+++ crates/core/src/codec/mod.rs
@@ -1,0 +1,3 @@
+pub mod anthropic;
+pub mod gemini;
+pub mod openai;
