+++ crates/provider/src/provider/openai.rs
@@ -1,11 +1,11 @@
 use crate::{ByteStream, HttpClient};
-use bytes::{Buf, Bytes, BytesMut};
+use bytes::Bytes;
 use crabllm_core::{
     AnthropicRequest, AnthropicResponse, AnthropicStreamEvent, AudioSpeechRequest, BoxStream,
     ChatCompletionChunk, ChatCompletionRequest, ChatCompletionResponse, EmbeddingRequest,
     EmbeddingResponse, Error, ImageRequest, MultipartField, Provider,
 };
-use futures::stream::{self, Stream, StreamExt};
+use futures::stream::{Stream, StreamExt};
 
 #[derive(Debug, Clone)]
 pub struct OpenaiProvider {
@@ -248,7 +248,7 @@ pub async fn chat_completion_stream(
     ];
     let byte_stream = client.post_stream(&url, &headers, body.into()).await?;
 
-    Ok(sse_stream(byte_stream))
+    Ok(crabllm_core::codec::openai::sse_stream(byte_stream))
 }
 
 /// Send an image generation request to an OpenAI-compatible endpoint.
@@ -343,63 +343,3 @@ pub async fn audio_transcription(
     Ok((resp.body, content_type))
 }
 
-/// Parse an SSE byte stream into `ChatCompletionChunk` items.
-pub(crate) fn sse_stream(
-    byte_stream: ByteStream,
-) -> impl Stream<Item = Result<ChatCompletionChunk, Error>> {
-    stream::unfold(
-        (byte_stream, BytesMut::new()),
-        |(mut byte_stream, mut buffer)| async move {
-            use futures::StreamExt;
-
-            loop {
-                if let Some(newline_pos) = buffer.iter().position(|&b| b == b'\n') {
-                    let mut line_end = newline_pos;
-                    if line_end > 0 && buffer[line_end - 1] == b'\r' {
-                        line_end -= 1;
-                    }
-                    let line = &buffer[..line_end];
-
-                    if line.is_empty() {
-                        buffer.advance(newline_pos + 1);
-                        continue;
-                    }
-
-                    if let Some(data) = line.strip_prefix(b"data: ") {
-                        let data = match std::str::from_utf8(data) {
-                            Ok(s) => s.trim(),
-                            Err(_) => {
-                                buffer.advance(newline_pos + 1);
-                                continue;
-                            }
-                        };
-                        if data == "[DONE]" {
-                            return None;
-                        }
-                        let result = match crabllm_core::json::from_str::<ChatCompletionChunk>(data)
-                        {
-                            Ok(chunk) => Ok(chunk),
-                            Err(e) => Err(Error::Decode(format!("SSE parse error: {e}"))),
-                        };
-                        buffer.advance(newline_pos + 1);
-                        return Some((result, (byte_stream, buffer)));
-                    }
-                    // Skip non-data lines (comments, event:, etc.)
-                    buffer.advance(newline_pos + 1);
-                    continue;
-                }
-
-                // Need more data from the stream.
-                match byte_stream.next().await {
-                    Some(Ok(bytes)) => {
-                        buffer.extend_from_slice(&bytes);
-                    }
-                    Some(Err(e)) => {
-                        return Some((Err(Error::Network(e.to_string())), (byte_stream, buffer)));
-                    }
-                    None => return None,
-                }
-            }
-        },
-    )
-}
